Changed the approach to static wrappers
Removed the old Glue file and cmd property
Created class properties for each class with static calls into it
Changed the constructor to initialize them with default production values, allowing a dependency injection for any of the classes while testing.
Had to dance a little, as PHP apparently doesn't allow class properties to be static class calls, but method variables can, so had to transfer the class property into a method variable before I could call the classes.
I relocated the test file and changed references to it. Using "suite" seemed overblown with just one file.
Reworked the constants for the test in the bootstrap file.
Added two glue files that will only be used during testing, not in production. They are the test doubles that get injected so we can trace the test behavior properly.
Revised both the "found" template and the default template to use the injectable static calls, and am now also testing the default template, instead of relying on the fact the "found" one is identical to it.